Add a KDUtils 0.1.10 recipe
Build KDUtils v0.1.10. This recipe assumes that Wayland will be part
of the image. In the future this could be improved to make the Wayland
depends optional based on the image configuration.

Add patches required to find whereami, find the correct Wayland
protocols XML files, and set library so versions for the KDUtils
libraries. These should eventually be upstreamed to KDUtils.